Impletest skip_fork_shutdown
When we are under php-fpm, master process goes inside startup hook once.
When our child ended/killed by some reason (max_requests for example), it visit shutdown hook.
But we don't need this! We only need to visit shutdown hook, when our master process stopped.
So we should check pid of current process in shutdown and allow shutdown only when it's same as was in startup.

* Document option skip_fork_shutdown
* Check for process id in shutdown when we enabled ini flag skip_fork_shutdown
* Test skip_fork_shutdown=0 with forks (current/default) and skip_fork_shutdown=1 (should be)
* Avoid compiler warnings (I'm a C-dummy, sorry K&R)